+++ src/shared/components/case-editor/services/cases.service.spec.ts
@@ -1,4 +1,4 @@
-import { Headers, Response, ResponseOptions } from '@angular/http';
+import { HttpHeaders, HttpResponse } from '@angular/common/http';
 import { Observable, throwError } from 'rxjs';
 
 import { AbstractAppConfig } from '../../../../app.config';
@@ -92,9 +92,7 @@ describe('CasesService', () => {
   describe('getCaseView()', () => {
 
     beforeEach(() => {
-      httpService.get.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        body: JSON.stringify(CASE_VIEW)
-      }))));
+      httpService.get.and.returnValue(Observable.of(CASE_VIEW));
     });
 
     it('should use HttpService::get with correct url', () => {
@@ -146,9 +144,7 @@ describe('CasesService', () => {
   describe('getCaseViewV2()', () => {
 
     beforeEach(() => {
-      httpService.get.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        body: JSON.stringify(CASE_VIEW)
-      }))));
+      httpService.get.and.returnValue(Observable.of(CASE_VIEW));
     });
 
     it('should use HttpService::get with correct url', () => {
@@ -157,10 +153,11 @@ describe('CasesService', () => {
         .subscribe();
 
       expect(httpService.get).toHaveBeenCalledWith(V2_CASE_VIEW_URL, {
-        headers: new Headers({
-          'Accept': CasesService.V2_MEDIATYPE_CASE_VIEW,
-          'experimental': 'true',
-        })
+        headers: new HttpHeaders()
+          .set('experimental', 'true')
+          .set('Accept', CasesService.V2_MEDIATYPE_CASE_VIEW)
+          .set('Content-Type', 'application/json'),
+        observe: 'body'
       });
     });
 
@@ -208,9 +205,7 @@ describe('CasesService', () => {
     const EVENT_TRIGGER: CaseEventTrigger = createCaseEventTrigger('', '', '', false, []);
 
     beforeEach(() => {
-      httpService.get.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        body: JSON.stringify(EVENT_TRIGGER)
-      }))));
+      httpService.get.and.returnValue(Observable.of(EVENT_TRIGGER));
     });
 
     it('should use HttpService::get with correct url for create case', () => {
@@ -218,11 +213,11 @@ describe('CasesService', () => {
         .getEventTrigger(CTID, EVENT_TRIGGER_ID, CASE_ID_UNDEFINED, 'true')
         .subscribe();
 
-      const headers = new Headers({
-        'experimental': 'true',
-        'Accept': CasesService.V2_MEDIATYPE_START_CASE_TRIGGER
-      });
-      expect(httpService.get).toHaveBeenCalledWith(EVENT_TRIGGER_FOR_CASE_TYPE_URL, {headers});
+      const headers = new HttpHeaders()
+        .set('experimental', 'true')
+        .set('Content-Type', 'application/json')
+        .set('Accept', CasesService.V2_MEDIATYPE_START_CASE_TRIGGER);
+      expect(httpService.get).toHaveBeenCalledWith(EVENT_TRIGGER_FOR_CASE_TYPE_URL, {headers, observe: 'body'});
     });
 
     it('should use HttpService::get with correct url for create event', () => {
@@ -230,11 +225,11 @@ describe('CasesService', () => {
         .getEventTrigger(CTID_UNDEFINED, EVENT_TRIGGER_ID, CASE_ID, 'true')
         .subscribe();
 
-      const headers = new Headers({
-        'experimental': 'true',
-        'Accept': CasesService.V2_MEDIATYPE_START_EVENT_TRIGGER
-      });
-      expect(httpService.get).toHaveBeenCalledWith(EVENT_TRIGGER_FOR_CASE_URL, {headers});
+      const headers = new HttpHeaders()
+        .set('experimental', 'true')
+        .set('Content-Type', 'application/json')
+        .set('Accept', CasesService.V2_MEDIATYPE_START_EVENT_TRIGGER);
+      expect(httpService.get).toHaveBeenCalledWith(EVENT_TRIGGER_FOR_CASE_URL, {headers, observe: 'body'});
     });
 
     it('should use HttpService::get with correct url for DRAFTS', () => {
@@ -242,11 +237,11 @@ describe('CasesService', () => {
         .getEventTrigger(CTID, EVENT_TRIGGER_ID, DRAFT_ID, 'true')
         .subscribe();
 
-      const headers = new Headers({
-          'experimental': 'true',
-          'Accept': CasesService.V2_MEDIATYPE_START_DRAFT_TRIGGER
-        });
-      expect(httpService.get).toHaveBeenCalledWith(EVENT_TRIGGER_DRAFT_URL, {headers});
+      const headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Content-Type', 'application/json')
+      .set('Accept', CasesService.V2_MEDIATYPE_START_DRAFT_TRIGGER);
+      expect(httpService.get).toHaveBeenCalledWith(EVENT_TRIGGER_DRAFT_URL, {headers, observe: 'body'});
     });
 
     it('should retrieve event trigger from server by case id', () => {
@@ -312,40 +307,38 @@ describe('CasesService', () => {
 
     const EVENT_RESPONSE = { id: 5 };
     const EMPTY_RESPONSE = { id: '' };
-    const HEADERS = new Headers({'content-type': CasesService.V2_MEDIATYPE_CREATE_EVENT});
+    const HEADERS = new HttpHeaders()
+      .set('content-type', CasesService.V2_MEDIATYPE_CREATE_EVENT);
 
     beforeEach(() => {
-      httpService.post.and.returnValue(Observable.of(new Response(new ResponseOptions({
+      httpService.post.and.returnValue(Observable.of({
         headers: HEADERS,
-        body: JSON.stringify(EVENT_RESPONSE)
-      }))));
+        body: EVENT_RESPONSE
+      }));
     });
 
     it('should use HttpService::post with correct url', () => {
       casesService
         .createEvent(CASE_DETAILS, CASE_EVENT_DATA)
         .subscribe();
+      const headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Accept', CasesService.V2_MEDIATYPE_CREATE_EVENT)
+      .set('Content-Type', 'application/json');
 
-      const headers = new Headers({
-        'experimental': 'true',
-        'Accept': CasesService.V2_MEDIATYPE_CREATE_EVENT
-      });
-
-      expect(httpService.post).toHaveBeenCalledWith(CREATE_EVENT_URL, CASE_EVENT_DATA, {headers});
+      expect(httpService.post).toHaveBeenCalledWith(CREATE_EVENT_URL, CASE_EVENT_DATA, {headers, observe: 'body'});
     });
 
     it('should create event on server', () => {
       casesService
         .createEvent(CASE_DETAILS, CASE_EVENT_DATA)
         .subscribe(
-          data => expect(data).toEqual(EVENT_RESPONSE)
+          data => expect((data as any).body).toEqual(EVENT_RESPONSE)
         );
     });
 
     it('should return body with empty id if no content-type response header', () => {
-      httpService.post.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        body: JSON.stringify(EVENT_RESPONSE)
-      }))));
+      httpService.post.and.returnValue(Observable.of(EVENT_RESPONSE));
 
       casesService
         .createEvent(CASE_DETAILS, CASE_EVENT_DATA)
@@ -402,9 +395,7 @@ describe('CasesService', () => {
     const EVENT_RESPONSE = { id: 5 };
 
     beforeEach(() => {
-      httpService.post.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        body: JSON.stringify(EVENT_RESPONSE)
-      }))));
+      httpService.post.and.returnValue(Observable.of(EVENT_RESPONSE));
     });
 
     it('should use HttpService::post with correct url', () => {
@@ -412,12 +403,12 @@ describe('CasesService', () => {
         .validateCase(CTID, CASE_EVENT_DATA, PAGE_ID)
         .subscribe();
 
-      const headers = new Headers({
-        'experimental': 'true',
-        'Accept': CasesService.V2_MEDIATYPE_CASE_DATA_VALIDATE
-      });
+      const headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Accept', CasesService.V2_MEDIATYPE_CASE_DATA_VALIDATE)
+      .set('Content-Type', 'application/json');
 
-      expect(httpService.post).toHaveBeenCalledWith(VALIDATE_CASE_URL, CASE_EVENT_DATA, {headers});
+      expect(httpService.post).toHaveBeenCalledWith(VALIDATE_CASE_URL, CASE_EVENT_DATA, {headers, observe: 'body'});
     });
 
     it('should validate case on server', () => {
@@ -455,13 +446,14 @@ describe('CasesService', () => {
 
     const CASE_RESPONSE = { id: 5 };
     const EMPTY_RESPONSE = { id: '' };
-    const HEADERS = new Headers({'content-type': 'application/json;charset=UTF-8'});
+    const HEADERS = new HttpHeaders()
+      .set('content-type', 'application/json;charset=UTF-8');
 
     beforeEach(() => {
-      httpService.post.and.returnValue(Observable.of(new Response(new ResponseOptions({
+      httpService.post.and.returnValue(Observable.of({
         headers: HEADERS,
-        body: JSON.stringify(CASE_RESPONSE)
-      }))));
+        body: CASE_RESPONSE
+      }));
     });
 
     it('should use HttpService::post with correct url', () => {
@@ -469,26 +461,24 @@ describe('CasesService', () => {
         .createCase(CTID, CASE_EVENT_DATA)
         .subscribe();
 
-      const headers = new Headers({
-        'experimental': 'true',
-        'Accept': CasesService.V2_MEDIATYPE_CREATE_CASE
-      });
+      const headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Accept', CasesService.V2_MEDIATYPE_CREATE_CASE)
+      .set('Content-Type', 'application/json');
 
-      expect(httpService.post).toHaveBeenCalledWith(CREATE_CASE_URL, CASE_EVENT_DATA, {headers});
+      expect(httpService.post).toHaveBeenCalledWith(CREATE_CASE_URL, CASE_EVENT_DATA, {headers, observe: 'body'});
     });
 
     it('should create case on server', () => {
       casesService
         .createCase(CTID, CASE_EVENT_DATA)
         .subscribe(
-          data => expect(data).toEqual(CASE_RESPONSE)
+          data => expect((data as any).body).toEqual(CASE_RESPONSE)
         );
     });
 
     it('should return body with empty id if no content-type response header', () => {
-      httpService.post.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        body: JSON.stringify(CASE_RESPONSE)
-      }))));
+      httpService.post.and.returnValue(Observable.of(CASE_RESPONSE));
 
       casesService
         .createCase(CTID, CASE_EVENT_DATA)
@@ -522,26 +512,25 @@ describe('CasesService', () => {
         }
       ]
     };
-    const HEADERS = new Headers({'content-type': CasesService.V2_MEDIATYPE_CASE_DOCUMENTS});
+
+    const HEADERS = new HttpHeaders()
+      .set('content-type', CasesService.V2_MEDIATYPE_CASE_DOCUMENTS);
 
     beforeEach(() => {
-      httpService.get.and.returnValue(Observable.of(new Response(new ResponseOptions({
-        headers: HEADERS,
-        body: JSON.stringify(DOCUMENTS)
-      }))));
+      httpService.get.and.returnValue(Observable.of(DOCUMENTS));
     });
 
     it('should use HttpService::get with correct url', () => {
-      const headers = new Headers({
-        'experimental': 'true',
-        'Accept': CasesService.V2_MEDIATYPE_CASE_DOCUMENTS
-      });
+      const headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Accept', CasesService.V2_MEDIATYPE_CASE_DOCUMENTS)
+      .set('Content-Type', 'application/json');
 
       casesService
         .getPrintDocuments(CASE_ID)
         .subscribe();
 
-      expect(httpService.get).toHaveBeenCalledWith(PRINT_DOCUMENTS_URL, {headers});
+      expect(httpService.get).toHaveBeenCalledWith(PRINT_DOCUMENTS_URL, {headers, observe: 'body'});
     });
 
     it('should retrieve document list from server', () => {