+++ src/shared/components/case-history/services/case-history.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { plainToClass } from 'class-transformer';
-import { Headers } from '@angular/http';
 import { HttpService, HttpErrorService } from '../../../services';
 import { AbstractAppConfig } from '../../../../app.config';
 import { CaseHistory } from '../domain';
+import { HttpHeaders } from '@angular/common/http';
 
 @Injectable()
 export class CaseHistoryService {
@@ -19,15 +19,14 @@ export class CaseHistoryService {
       eventId: string): Observable<CaseHistory> {
 
     const url = this.appConfig.getCaseHistoryUrl(caseId, eventId);
-
-    let headers = new Headers({
-      'experimental': 'true',
-      'Accept': CaseHistoryService.V2_MEDIATYPE_CASE_EVENT_VIEW
-    });
+    let headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Accept', CaseHistoryService.V2_MEDIATYPE_CASE_EVENT_VIEW)
+      .set('Content-Type', 'application/json');
 
     return this.httpService
-      .get(url, {headers})
-      .map(response => response.json())
+      .get(url, {headers, observe: 'body'})
+      .map(response => response)
       .catch((error: any): any => {
         this.httpErrorService.setError(error);
         return Observable.throw(error);