+++ src/shared/services/search/search.service.ts
@@ -1,14 +1,13 @@
 import { Injectable } from '@angular/core';
-import { RequestOptionsArgs } from '@angular/http';
 import { finalize, map } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { AbstractAppConfig } from '../../../app.config';
-import { HttpService } from '../http';
-import { Headers } from '@angular/http';
+import { HttpService, OptionsType } from '../http';
 import { RequestOptionsBuilder, SearchView } from '../request';
 import { SearchInput } from '../../components/search-filters';
 import { SearchResultView } from '../../domain/search';
 import { LoadingService } from '../loading';
+import { HttpHeaders } from '@angular/common/http';
 
 @Injectable()
 export class SearchService {
@@ -32,12 +31,12 @@ export class SearchService {
                                            + `/case-types/${caseTypeId}`
                                            + `/cases`;
 
-    let options: RequestOptionsArgs = this.requestOptionsBuilder.buildOptions(metaCriteria, caseCriteria, view);
+    let options: OptionsType  = this.requestOptionsBuilder.buildOptions(metaCriteria, caseCriteria, view);
     const loadingToken = this.loadingService.register();
     return this.httpService
       .get(url, options)
       .pipe(
-        map(response => response.json()),
+        map(response => response),
         finalize(() => this.loadingService.unregister(loadingToken))
       );
   }
@@ -46,7 +45,7 @@ export class SearchService {
                 metaCriteria: object, caseCriteria: object, view?: SearchView, sort?: {column: string, order: number}): Observable<{}> {
     const url = this.appConfig.getCaseDataUrl() + `/internal/searchCases?ctid=${caseTypeId}&use_case=${view}`;
 
-    let options: RequestOptionsArgs = this.requestOptionsBuilder.buildOptions(metaCriteria, caseCriteria, view);
+    let options: OptionsType = this.requestOptionsBuilder.buildOptions(metaCriteria, caseCriteria, view);
     const body: {} = {
       sort,
       size: this.appConfig.getPaginationPageSize()
@@ -55,7 +54,7 @@ export class SearchService {
     return this.httpService
       .post(url, body, options)
       .pipe(
-        map(response => response.json()),
+        map(response => response),
         finalize(() => this.loadingService.unregister(loadingToken))
       );
   }
@@ -66,17 +65,17 @@ export class SearchService {
 
   getSearchInputs(jurisdictionId: string, caseTypeId: string): Observable<SearchInput[]> {
     let url = this.getSearchInputUrl(caseTypeId);
-    const headers = new Headers({
-      'Accept': SearchService.V2_MEDIATYPE_SEARCH_INPUTS,
-      'experimental': 'true',
-    });
+    const headers = new HttpHeaders()
+      .set('experimental', 'true')
+      .set('Accept', SearchService.V2_MEDIATYPE_SEARCH_INPUTS)
+      .set('Content-Type', 'application/json');
     this.currentJurisdiction = jurisdictionId;
     this.currentCaseType = caseTypeId;
     return this.httpService
-      .get(url, { headers })
+      .get(url, { headers, observe: 'body' })
       .pipe(
         map(response => {
-          let jsonResponse = response.json();
+          let jsonResponse = response;
           let searchInputs = jsonResponse.searchInputs;
           if (this.isDataValid(jurisdictionId, caseTypeId)) {
             searchInputs.forEach(item => {