+++ RELEASE-NOTES.md
@@ -1,4 +1,18 @@
 ## RELEASE NOTES
+### Version 2.72.5-remove-null-http-request-headers
+**EUI-3464** Fix Document field uploads by removing setting of null HTTP "Accept" and "Content-Type" headers
+
+### Version 2.72.4-check-your-answers
+Fix EUI-3452 - fix for show/hide conditions on Check Your Answers page.
+
+### Version 2.72.3-dynamic-list-null-values
+EUI-3045 Handle null values appropriately.
+
+### Version 2.72.2
+EUI-3045 performance enhancements and 3045 work
+
+### Version 2.72.1-do-not-send-empty-documents
+Fix EUI-3403 - fixed a subsequent issue with empty documents being sent.
 
 ### Version 2.72.0-task-event-completion
 **EUI-3229** Task event completion - part 2